Split large files to smaller ones
